Fix - usage of not existing $this->db
In this class there was a mix of the usage of $this->db and $db from global. As the class and the parent class doesn't have a db var, switch to usage everywhere from global db var. Without this it crashes on dolibarr version 22 for adding a new line for example in an custom invoice.